+++ module/PassGen2/examples/Update-PassGen2.ps1
@@ -1,7 +1,7 @@
 [CmdletBinding()]
 param(
     [Parameter()]
-    [string]$ModuleName = 'PassGen'
+    [string]$ModuleName = 'PassGen2'
 )
 
 $installed = Get-Module -ListAvailable -Name $ModuleName | Sort-Object Version -Descending | Select-Object -First 1